AFT setup to use a file for data
